StockExchange/microservices/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
aiosmtplib==3.0.1
prometheus-client==0.19.0
structlog==23.2.0
psycopg2-binary==2.9.9
influxdb-client==1.38.0
# OpenTelemetry tracing dependencies
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-asyncpg==0.42b0
opentelemetry-instrumentation-redis==0.42b0